Reorganize folders and directory - Move global.css to style directory - Create a directory called element - Move JS elements to element directory
