+++ src/elements/linked_list_element.js
